+++ src/arch/linux/x86_64/epoll.h
@@ -5,53 +5,53 @@
 #include "../errno.h"
 #include "./asm.h"
 
-static inline long linux_x8664_epoll_ctl(
+static inline int32_t linux_x8664_epoll_ctl(
     const int32_t        epoll_fd,
     const int32_t        op,
     const int32_t        fd,
     PWebSocketEpollEvent event)
 {
-    long ret = linux_x8664_asm_syscall4(
+    int32_t ret = linux_x8664_asm_syscall4(
         __NR_epoll_ctl,
         epoll_fd,
         op,
         fd,
         event);
 
-    if ((unsigned long)ret >= (unsigned long)-4095) {
+    if ((uint64_t)ret >= (uint64_t)-4095) {
         errno = -ret;
         ret   = -1;
     }
     return ret;
 }
 
-static inline long linux_x8664_epoll_create1(const int32_t flags)
+static inline int32_t linux_x8664_epoll_create1(const int32_t flags)
 {
-    long ret = linux_x8664_asm_syscall1(
+    int32_t ret = linux_x8664_asm_syscall1(
         __NR_epoll_create1,
         flags);
 
-    if ((unsigned long)ret >= (unsigned long)-4095) {
+    if ((uint64_t)ret >= (uint64_t)-4095) {
         errno = -ret;
         ret   = -1;
     }
     return ret;
 }
 
-static inline long linux_x8664_epoll_wait(
+static inline int32_t linux_x8664_epoll_wait(
     const int32_t        epfd,
     PWebSocketEpollEvent events,
     const int32_t        maxevents,
     const int32_t        timeout)
 {
-    long ret = linux_x8664_asm_syscall4(
+    int32_t ret = linux_x8664_asm_syscall4(
         __NR_epoll_wait,
         epfd,
         events,
         maxevents,
         timeout);
 
-    if ((unsigned long)ret >= (unsigned long)-4095) {
+    if ((uint64_t)ret >= (uint64_t)-4095) {
         errno = -ret;
         ret   = -1;
     }
